Refactor type hints and casting
This commit introduces several type hint improvements and uses `cast`
where appropriate to enhance code clarity and maintainability.

Specifically, it addresses:
- Type hinting for `_DATABASE` in `OpticalEquipment`.
- Casting of `_DATABASE` values in
  `OpticalEquipment.get_database_entries`.
- Casting of `CoordinateSystem` enum values in plotting tests.
- Casting of ephemeris object in `skyfield_searches`.
- Ensures `n_subs` is an integer in `OpticalPath`.
- Added assertions for `None` checks in several test cases.

+++ apts/optics.py
@@ -644,7 +644,7 @@ class OpticalPath:
         """
         low = 0.0
         high = 30.0
-        n_subs = float(numpy.ceil(total_integration_time / sub_exposure_time))
+        n_subs = int(numpy.ceil(total_integration_time / sub_exposure_time))
 
         for _ in range(12):  # 2^12 = 4096, plenty for 0.01 prec in 30 range
             mid = (low + high) / 2
